move class to different package
