+++ core/src/main/java/me/approximations/aMessaging/bungee/message/actions/normal/ConnectOtherAction.java
@@ -22,22 +22,22 @@
  * SOFTWARE.
  */
 
-package me.approximations.aMessaging.bungee.message.actions;
+package me.approximations.aMessaging.bungee.message.actions.normal;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import me.approximations.aMessaging.bungee.message.response.handler.MessageResponseHandler;
+import me.approximations.aMessaging.bungee.message.actions.MessageAction;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.DataOutput;
 import java.io.IOException;
-import java.util.concurrent.CompletableFuture;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
-public class GetPlayerServerAction extends ResponseableMessageAction<String, String> {
-    public static final String SUB_CHANNEL = "GetPlayerServer";
+public class ConnectOtherAction extends MessageAction {
+    public static final String SUB_CHANNEL = "ConnectOther";
     private final String playerName;
+    private final String serverName;
 
     @Override
     public @NotNull String getSubChannel() {
@@ -45,17 +45,9 @@ public class GetPlayerServerAction extends ResponseableMessageAction<String, Str
     }
 
     @Override
-    public void writeHead(@NotNull DataOutput out) throws IOException {
-        out.writeUTF(SUB_CHANNEL);
-        out.writeUTF(playerName);
-    }
-
-    @Override
-    public @NotNull CompletableFuture<String> addFuture(MessageResponseHandler<String, String> responseHandler) {
-        final CompletableFuture<String> future = new CompletableFuture<>();
-
-        responseHandler.addFuture(playerName, future);
-
-        return future;
+    public void writeHead(@NotNull DataOutput dataOutput) throws IOException {
+        dataOutput.writeUTF(SUB_CHANNEL);
+        dataOutput.writeUTF(playerName);
+        dataOutput.writeUTF(serverName);
     }
 }
