+++ core/src/main/java/me/approximations/aMessaging/bungee/message/actions/responseable/GetPlayerServerAction.java
@@ -22,22 +22,23 @@
  * SOFTWARE.
  */
 
-package me.approximations.aMessaging.bungee.message.actions;
+package me.approximations.aMessaging.bungee.message.actions.responseable;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import me.approximations.aMessaging.bungee.message.actions.ResponseableMessageAction;
 import me.approximations.aMessaging.bungee.message.response.handler.MessageResponseHandler;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.DataOutput;
 import java.io.IOException;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
-public class GetServersAction extends ResponseableMessageAction<String, List<String>> {
-    public static final String SUB_CHANNEL = "GetServers";
+public class GetPlayerServerAction extends ResponseableMessageAction<String, String> {
+    public static final String SUB_CHANNEL = "GetPlayerServer";
+    private final String playerName;
 
     @Override
     public @NotNull String getSubChannel() {
@@ -47,13 +48,14 @@ public class GetServersAction extends ResponseableMessageAction<String, List<Str
     @Override
     public void writeHead(@NotNull DataOutput out) throws IOException {
         out.writeUTF(SUB_CHANNEL);
+        out.writeUTF(playerName);
     }
 
     @Override
-    public @NotNull CompletableFuture<List<String>> addFuture(MessageResponseHandler<String, List<String>> responseHandler) {
-        final CompletableFuture<List<String>> future = new CompletableFuture<>();
+    public @NotNull CompletableFuture<String> addFuture(MessageResponseHandler<String, String> responseHandler) {
+        final CompletableFuture<String> future = new CompletableFuture<>();
 
-        responseHandler.addFuture(null, future);
+        responseHandler.addFuture(playerName, future);
 
         return future;
     }
