+++ core/src/main/java/me/approximations/aMessaging/bungee/message/actions/responseable/GetServerAction.java
@@ -22,23 +22,22 @@
  * SOFTWARE.
  */
 
-package me.approximations.aMessaging.bungee.message.actions;
+package me.approximations.aMessaging.bungee.message.actions.responseable;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import me.approximations.aMessaging.bungee.message.actions.ResponseableMessageAction;
 import me.approximations.aMessaging.bungee.message.response.handler.MessageResponseHandler;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.DataOutput;
 import java.io.IOException;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
-public class PlayerListAction extends ResponseableMessageAction<String, List<String>> {
-    public static final String SUB_CHANNEL = "PlayerList";
-    private final String server;
+public class GetServerAction extends ResponseableMessageAction<Void, String> {
+    public static final String SUB_CHANNEL = "GetServer";
 
     @Override
     public @NotNull String getSubChannel() {
@@ -48,14 +47,13 @@ public class PlayerListAction extends ResponseableMessageAction<String, List<Str
     @Override
     public void writeHead(@NotNull DataOutput out) throws IOException {
         out.writeUTF(SUB_CHANNEL);
-        out.writeUTF(server);
     }
 
     @Override
-    public @NotNull CompletableFuture<List<String>> addFuture(MessageResponseHandler<String, List<String>> responseHandler) {
-        final CompletableFuture<List<String>> future = new CompletableFuture<>();
+    public @NotNull CompletableFuture<String> addFuture(MessageResponseHandler<Void, String> responseHandler) {
+        final CompletableFuture<String> future = new CompletableFuture<>();
 
-        responseHandler.addFuture(server, future);
+        responseHandler.addFuture(null, future);
 
         return future;
     }
