+++ core/src/main/java/me/approximations/aMessaging/bungee/message/actions/responseable/PlayerListAction.java
@@ -22,21 +22,23 @@
  * SOFTWARE.
  */
 
-package me.approximations.aMessaging.bungee.message.actions;
+package me.approximations.aMessaging.bungee.message.actions.responseable;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import me.approximations.aMessaging.bungee.message.actions.ResponseableMessageAction;
 import me.approximations.aMessaging.bungee.message.response.handler.MessageResponseHandler;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.DataOutput;
 import java.io.IOException;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 @EqualsAndHashCode(callSuper = true)
 @Data
-public class PlayerCountAction extends ResponseableMessageAction<String, Integer> {
-    public static final String SUB_CHANNEL = "PlayerCount";
+public class PlayerListAction extends ResponseableMessageAction<String, List<String>> {
+    public static final String SUB_CHANNEL = "PlayerList";
     private final String server;
 
     @Override
@@ -51,8 +53,8 @@ public class PlayerCountAction extends ResponseableMessageAction<String, Integer
     }
 
     @Override
-    public @NotNull CompletableFuture<Integer> addFuture(MessageResponseHandler<String, Integer> responseHandler) {
-        final CompletableFuture<Integer> future = new CompletableFuture<>();
+    public @NotNull CompletableFuture<List<String>> addFuture(MessageResponseHandler<String, List<String>> responseHandler) {
+        final CompletableFuture<List<String>> future = new CompletableFuture<>();
 
         responseHandler.addFuture(server, future);
